move polyamorous from get to a module
