+++ lib/polyamorous/activerecord_5.2_ruby_2/reflection.rb
