+++ lib/polyamorous/activerecord_6.0_ruby_2/join_association.rb
@@ -1,0 +1,1 @@
+require 'polyamorous/activerecord_5.2_ruby_2/join_association'
