+++ lib/polyamorous/activerecord_6.0_ruby_2/join_dependency.rb
@@ -1,3 +1,4 @@
+# active_record_6.0_ruby_2/join_dependency.rb
 module Polyamorous
   module JoinDependencyExtensions
     # Replaces ActiveRecord::Associations::JoinDependency#build
@@ -27,7 +28,7 @@ module Polyamorous
       end
     end
 
-    def join_constraints(joins_to_add, join_type, alias_tracker)
+    def join_constraints(joins_to_add, alias_tracker)
       @alias_tracker = alias_tracker
 
       construct_tables!(join_root)
@@ -36,9 +37,9 @@ module Polyamorous
       joins.concat joins_to_add.flat_map { |oj|
         construct_tables!(oj.join_root)
         if join_root.match?(oj.join_root) && join_root.table.name == oj.join_root.table.name
-          walk join_root, oj.join_root
+          walk join_root, oj.join_root, oj.join_type
         else
-          make_join_constraints(oj.join_root, join_type)
+          make_join_constraints(oj.join_root, oj.join_type)
         end
       }
     end
