+++ lib/polyamorous/activerecord_6.1_ruby_2/join_dependency.rb
@@ -1,0 +1,1 @@
+require 'polyamorous/activerecord_6.0_ruby_2/join_dependency'
