+++ lib/polyamorous/tree_node.rb
