+++ lib/ransack/adapters/active_record.rb
@@ -1,4 +1,7 @@
 require 'ransack/adapters/active_record/base'
+# require 'polyamorous/polyamorous.rb'
+
+
 
 ActiveSupport.on_load(:active_record) do
   extend Ransack::Adapters::ActiveRecord::Base
