+++ lib/ransack/context.rb
@@ -1,3 +1,4 @@
+# require 'polyamorous/polyamorous.rb'
 require 'ransack/visitor'
 Ransack::Adapters.object_mapper.require_context
