--- polyamorous/lib/polyamorous/activerecord_6.0_ruby_2/join_association.rb
@@ -1,2 +1,0 @@
-# active_record_6.0_ruby_2/join_association
-require 'polyamorous/activerecord_5.2_ruby_2/join_association'
