--- polyamorous/lib/polyamorous/activerecord_6.1_ruby_2/join_dependency.rb
@@ -1,2 +1,0 @@
-# active_record_6.1_ruby_2/join_dependency.rb
-require 'polyamorous/activerecord_6.0_ruby_2/join_dependency'
