--- polyamorous/lib/polyamorous/version.rb
@@ -1,3 +1,0 @@
-module Polyamorous
-  VERSION = '2.3.2'
-end
