+++ ransack.gemspec
@@ -16,11 +16,11 @@ Gem::Specification.new do |s|
 
   s.add_dependency 'activerecord', '>= 5.2.1'
   s.add_dependency 'activesupport', '>= 5.2.1'
+  s.add_dependency 'machinist', '~> 1.0.6'
   s.add_dependency 'i18n'
-  s.add_dependency 'polyamorous', '2.3.2'
 
-  s.files         = `git ls-files`.split("\n").reject { |f| f.match(%r{^(polyamorous/|logo/|coverage/|spec/)}) }
-  s.test_files    = `git ls-files -- spec/*`.split("\n")
-
+  s.files         = `git ls-files`.split("\n")
+  s.test_files    = `git ls-files -- {test,spec,features}/*`.split("\n")
+  s.executables   = `git ls-files -- bin/*`.split("\n").map{ |f| File.basename(f) }
   s.require_paths = ["lib"]
 end
